requirements.txt
Flask==2.1.0
cryptography==3.4.7
python-dotenv==0.19.0
pytest==5.3.5
pytest-cov==2.8.1